feat: 詳細ページをCMSと結合する
