+++ src/app/project/[projectId]/page.tsx
@@ -1,18 +1,19 @@
 import { DetailProject } from '@/components/templates/detail-project/DetailProject';
-import { Project } from '@/types/project';
+import { getProject } from '@/service/get-project-service';
+import { notFound } from 'next/navigation';
 
-const project: Project = {
-  id: 'dummy1',
-  title: 'dummy1',
-  description: 'dummy1',
-  tags: ['tag1', 'tag2'],
-  thumbnails: ['/dummy.png', '/dummy.png'],
-  resources: [],
-  recipe: '',
-};
-
-const DetailProjectPage = ({ params }: { params: { projectId: string } }) => {
+const DetailProjectPage = async ({
+  params,
+}: {
+  params: { projectId: string };
+}) => {
   console.info(`Detail: ${params.projectId}`);
+
+  const project = await getProject(params.projectId);
+
+  if (!project) {
+    return notFound();
+  }
 
   return <DetailProject project={project} />;
 };
